Fix: num.clamp() no es int/double, rompía el build en 10 sitios
El build falló en cropToGuideSquare (detection_crop.dart:73): num.clamp()
está declarado en la clase num y devuelve num, no el tipo del receptor —
un int.clamp(int, int) no da int, da num. Sin un .toInt()/.toDouble()
explícito después, el resultado no encaja donde el contexto exige un tipo
concreto (ancho/alto de img.copyCrop, índice de lista, Duration.milliseconds,
elemento de Uint8List/Int8List, etc.).

Encontrado y corregido en los 3 sitios nuevos de este fix (cropToGuideSquare,
y los dos bucket-level del watershed en segmentForeground) y además en 7
sitios preexistentes con la misma forma, que iban a fallar uno por uno en
builds sucesivos apenas se corrigiera el primero: centerCrop y los índices
de _sobelX/_sobelY en detection_crop.dart, el progreso de similitud en
real_search_screen.dart, ttsRate en app_settings.dart, la cuantización INT8
en embedding_service_native.dart, el intervalo adaptativo y la conversión
YUV420→RGB en camera_detection_view.dart, y el recorte de líneas en
metrics_logger.dart.

Sin Flutter SDK en el entorno: no se pudo compilar acá, sigue haciendo
falta un build real en el teléfono para confirmar que no queda ninguno.

+++ aura_app/lib/screens/camera_detection_view.dart
@@ -384,7 +384,7 @@ class _CameraDetectionViewState extends State<CameraDetectionView>
       if (_slowFrameCount >= 3) {
         final newMs = (_adaptiveFrameInterval.inMilliseconds + 100).clamp(
           _minFrameInterval.inMilliseconds, 500,
-        );
+        ).toInt();
         _adaptiveFrameInterval = Duration(milliseconds: newMs);
         _slowFrameCount = 0;
       }
@@ -394,7 +394,7 @@ class _CameraDetectionViewState extends State<CameraDetectionView>
       if (_adaptiveFrameInterval > _minFrameInterval) {
         final newMs = (_adaptiveFrameInterval.inMilliseconds - 50).clamp(
           _minFrameInterval.inMilliseconds, 500,
-        );
+        ).toInt();
         _adaptiveFrameInterval = Duration(milliseconds: newMs);
       }
     }
@@ -1129,9 +1129,9 @@ Uint8List _yuv420ToRgb(_YuvConversionArgs a) {
       final uvIdx = (y >> 1) * a.uvStride + (x >> 1) * a.uvPixelStride;
       final uu = a.uBytes[uvIdx] - 128;
       final vv = a.vBytes[uvIdx] - 128;
-      rgb[idx++] = ((298 * yy + 409 * vv + 128) >> 8).clamp(0, 255);
-      rgb[idx++] = ((298 * yy - 100 * uu - 208 * vv + 128) >> 8).clamp(0, 255);
-      rgb[idx++] = ((298 * yy + 516 * uu + 128) >> 8).clamp(0, 255);
+      rgb[idx++] = ((298 * yy + 409 * vv + 128) >> 8).clamp(0, 255).toInt();
+      rgb[idx++] = ((298 * yy - 100 * uu - 208 * vv + 128) >> 8).clamp(0, 255).toInt();
+      rgb[idx++] = ((298 * yy + 516 * uu + 128) >> 8).clamp(0, 255).toInt();
     }
   }
   return rgb;
